Index Refactor (#65)
Our index.tsx file was unnecessarily complicated. There were also a lot of state variables that we did not have to manage. This PR attempts to correct all such issues. What has changed is:

General reduction of props drilling so that our interfaces and components can be simpler.
Simplification of props and variable names.
More helper functions and utils to simplify components
UseContext for user(self).
New EnhancedPublicUser type that contains information on whether a user is favorited and any incoming or outgoing requests available on them.
Already-Connected Modal has been removed in favor for a toast notification. This makes things a lot easier to deal with when it comes to drilling down setStates functions for one-off uses.
Toast notifications in top-right (user test feedback)

Other stuff that I don't remember off the top of my head, but the diff is pretty clear, or contact @devashish9 for any additional questions.